Eliminated a LOT of errors... hopefully this will stabilise runs. Valgrind reports no leaks or errors except pool errors in libgomp, at least for FMMTest

+++ include/branch.hpp
@@ -38,7 +38,7 @@ public:
 
     Branch(Node *parent, int i, int j, int k);
     
-    static unsigned long int NumBranches;
+    static int NumBranches;
 
     static Array < Array <int> > srcMomentInds, trgMomentInds, srcInheritInds, trgInheritInds;
